Move tasks copying to new post-install script
