+++ package.json
@@ -4,7 +4,7 @@
   "description": "hacka-news for the web",
   "main": "app.js",
   "scripts": {
-    "postinstall": "cd semantic && gulp build && cd .. && npm run dl-vendors && npm run build",
+    "postinstall": "node post-install.js && cd semantic && gulp build && cd .. && npm run dl-vendors && npm run build",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node app.js",
     "build": "webpack",
