+++ semantic/gulpfile.js
@@ -3,21 +3,6 @@
 *******************************/
 
 var
-    fs           = require('fs-extra'),
-    path         = require('path');
-
-/* Copy tasks from node_modules */
-
-try {
-    var stat = fs.statSync(path.join(__dirname, './tasks'));
-    console.log('"tasks" folder already exists. Continuing...');
-} catch (e) {
-    console.log('Copying "tasks" folder from "node_modules/semantic-ui/tasks"');
-    fs.copySync(path.join(__dirname, '../node_modules/semantic-ui/tasks'), path.join(__dirname, './tasks'));
-    console.log('Copying done! Continuing...');
-}
-
-var
   gulp         = require('gulp-help')(require('gulp')),
 
   // read user config to know what task to load
@@ -44,6 +29,7 @@ var
   buildRTL     = require('./tasks/rtl/build'),
   watchRTL     = require('./tasks/rtl/watch')
 ;
+
 
 /*******************************
              Tasks
